Moxfy_adminPanel/requirements.txt
flet==0.12.1
pymongo==4.5.0
python-dotenv==1.0.0
